move sudo to package
